Extend duplicates error message
