+++ src/upgini/resource_bundle/strings.properties
@@ -114,7 +114,8 @@ dataset_too_many_rows_unregistered=For unregistered users total rows limit for X
 dataset_empty_column_names=Some column names are empty. Add names please
 dataset_too_long_column_name=Column {} is too long: {} characters. Remove this column or trim length to 50 characters
 dataset_full_duplicates=WARNING: {:.5f}% of the rows are fully duplicated
-dataset_diff_target_duplicates={:.4f}% of rows ({}) in X are duplicates with different y values. Please check X dataframe
+dataset_diff_target_duplicates={:.4f}% of rows ({}) in X and eval_set are duplicates with different y values. Please check dataframes
+duplicates_sample=Duplicates sample:
 dataset_drop_old_dates=WARNING: We don't have data before '2000-01-01' and removed all earlier records from the search dataset
 dataset_all_dates_old=There is empty train dataset after removing data before '2000-01-01'
 dataset_invalid_target_type=Unexpected dtype of target for binary task type: {}. Expected int or bool
